chore: Manual JSX in with-pubsub-hoc
